Simran - Cancel booking | DB changes | Room changes

+++ hotel-booking-service/src/main/java/com/hotel/sjsu/hotelbookingservice/model/Hotel.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Component;
 @Component
 public class Hotel {
 
-    private Integer hotelId;
+    private Long hotelId;
     private String hotelname;
     private String hotelLocation;
     private String hotel_address;
